requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-annoying==0.10.6
django-boost==2.1
django-embed-video==1.4.9
idna==3.4
Pillow==10.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
ua-parser==0.18.0
urllib3==2.0.4
user-agents==2.2.0
